fcsc_Server/requirements.txt
google-generativeai
openai
pandas 
python-dotenv 
matplotlib 
seaborn 
sqlparse
pandasql
openpyxl
sqlalchemy
psycopg2-binary
supabase
fastapi 
uvicorn
plotly
reportlab
markdown
googletrans
arabic_reshaper
python-bidi
bs4
python-multipart